AI/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
duckdb
pandas
plotly
groq
langgraph
langchain
langchain-groq
langsmith
supabase
